requirements.txt
# absl-py==2.3.1
# appnope==0.1.4
# asttokens==3.0.0
# astunparse==1.6.3
# attrs==25.4.0
# certifi==2025.10.5
# charset-normalizer==3.4.3
# comm==0.2.3
# contourpy==1.3.3
# cycler==0.12.1
# debugpy==1.8.17
# decorator==5.2.1
# dm-tree==0.1.9
# docstring_parser==0.17.0
# einops==0.8.1
# etils==1.13.0
# executing==2.2.1
# filetype==1.2.0
# flatbuffers==25.9.23
# fonttools==4.60.1
# fsspec==2025.9.0
# gast==0.6.0
# google-pasta==0.2.0
# googleapis-common-protos==1.70.0
# grpcio==1.75.1
# h5py==3.14.0
# idna==3.7
# immutabledict==4.2.2
# importlib_resources==6.5.2
# ipykernel==6.30.1
# ipython==9.6.0
# ipython_pygments_lexers==1.1.1
# jedi==0.19.2
# jupyter_client==8.6.3
# jupyter_core==5.8.1
# keras==3.11.3
# kiwisolver==1.4.9
# libclang==18.1.1
# Markdown==3.9
# markdown-it-py==4.0.0
# MarkupSafe==3.0.3
# matplotlib==3.10.7
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# ml_dtypes==0.5.3
# namex==0.1.0
# nest-asyncio==1.6.0
# numpy==2.3.3
# opencv-python-headless==4.10.0.84
# opt_einsum==3.4.0
# optree==0.17.0
# packaging==25.0
# parso==0.8.5
# pexpect==4.9.0
# pi_heif==1.1.1
# pillow==11.3.0
# pillow-avif-plugin==1.5.2
# platformdirs==4.5.0
# promise==2.3
# prompt_toolkit==3.0.52
# protobuf==6.32.1
# psutil==7.1.0
# ptyprocess==0.7.0
# pure_eval==0.2.3
# pyarrow==21.0.0
# Pygments==2.19.2
# pyparsing==3.2.5
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.1
# PyYAML==6.0.3
# pyzmq==27.1.0
# requests==2.32.5
# requests-toolbelt==1.0.0
# rich==14.2.0
# roboflow==1.2.10
# simple-parsing==0.1.7
# six==1.17.0
# stack-data==0.6.3
# tensorboard==2.20.0
# tensorboard-data-server==0.7.2
# tensorflow==2.20.0
# tensorflow-datasets==4.9.9
# tensorflow-hub==0.16.1
# tensorflow-metadata==1.17.2
# termcolor==3.1.0
# tf_keras==2.20.1
# toml==0.10.2
# tornado==6.5.2
# tqdm==4.67.1
# traitlets==5.14.3
# typing_extensions==4.15.0
# urllib3==2.5.0
# wcwidth==0.2.14
# Werkzeug==3.1.3
# wrapt==1.17.3
# zipp==3.23.0
# fastapi==0.115.*
# uvicorn[standard]==0.30.*
# pillow==10.*
# numpy==1.26.*
# tensorflow-cpu==2.15.*


fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.9     # needed for UploadFile form parsing
pillow==10.4.0              # image decoding
numpy              # TF 2.15 expects numpy<2
h5py==3.14.0               # load .weights.h5
protobuf<4.25.0             # safer with TF 2.15 wheels
tensorflow     # CPU build, works on python:3.11-slim
psutil


